requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
googletrans==3.1.0a0
requests==2.31.0
tqdm==4.66.1
langdetect==1.0.9
openpyxl==3.1.2
pandas==2.1.4